clould credentials and private key
Move clouds.yml to the top level directory: there is no reason to have
it in the groups_vars/all directory.

Remove the ssh_public_keyfile variable and assume it can be generated
by suffixing .pub

Rename the secrets.yml file to private-key.yml.example because it is
more specific. Also add a symlink in bootstrap to create
private-key.yml as a symlink to private-key.yml.example in test
environments. In production environment the file will be manually
created to point to the actual private key to be used.