packages: fix to work in preprod
a) Do not use docker cache when building

If the test is run more than once the docker cache may become a
problem. For instance the apt-get update run a few days ago may no
longer be accurate because some packages are not found anymore.

Run apt-get install on the same line as apt-get update so they
belong to the same layer.

Add --no-cache so that indirect dependencies of the ossec package used
to test installation are not impacted.

b) Load letsencrypt staging certificates

When running in preproduction the certificates are not valid and need
to be explicitly loaded or ignored.

Fixes main/securedrop-club#52