backup: packages-host contains signing keys and old packages
We may not be able to rebuild old packages as we do not maintain
a stable development environment that matches the stable packages that
are released. However, we want to keep the old packages for
posterity. Chances are it will not be possible to setup SecureDrop
with these packages because it has external dependencies that are
either gone or modified.