bind: use ssh-keyscan to generate SSHFP records
The install_ssh_records role is re-architectured to operate on a
single host / port (i.e. install_ssh_records_host and
install_ssh_records_port ) pair given in argument. It is called from
the bind-client playbook for all known hosts to make sure SSHFP
records exists for all known hosts. But it can also be used to add
SSHFP records for a host / port that is not known in the inventory.

The zone file is modified to always $INCLUDE a file named after each
host. These files are initially empty and it is the responsibility of
ech ohost to populate them. Having a separate zone file for each host
also addresses a race: if N hosts try to update the zone file via a
delegate_to bind-host, they will override each other because there is
no lock in the blockinfile role.

If the install_ssh_records_host is not in the inventory, the SSHFP
records are included in the main zone file of the domain because there
is no way to figure out which $INCLUDE files is relevant.

The openssh-client package is from buster because the ssh-keyscan -D
option does not exist in openssh-client available in stretch or
stretch-backports.