bind: do not always change the master & zone files
The zone was always set and the bind server restarted. It was fine
when the zone was entirely defined by the template. But it now is
defined by multiple nsupdate. While this is happening the server
temporarily regresses to a state where it knows no domain.

The zone and master files are now changed only when they do not
already exist.

Fixes: main/infrastructure#74
Fixes: main/infrastructure#47